Refactor: Renames middlewares folder to middleware in lib and tests
