+++ lib/index.ts
@@ -1,5 +1,5 @@
 export type { IHttpRequest, IHttpResponse } from "./types";
 export { environmentConfiguration } from "./config";
 export type { ISwiftApplication } from "./application";
-export { jsonMiddleware } from "./middlewares";
+export { jsonMiddleware } from "./middleware";
 export { SwiftFactory } from "./factory";
