--- lib/middlewares/index.ts
@@ -1,4 +1,0 @@
-export type { IMiddlewareManager } from "./middleware-manager.interface";
-export type { Middleware } from "./middleware.types";
-export { MiddlewareManager } from "./middleware-manager";
-export { jsonMiddleware } from "./parsing/json-middleware";
