--- lib/middlewares/middleware-manager.interface.ts
@@ -1,11 +1,0 @@
-import { TMiddlewareFunction } from "../middlewares";
-import { IHttpRequest, IHttpResponse } from "../types";
-
-export interface IMiddlewareManager {
-  use(middleware: TMiddlewareFunction): void;
-  executeMiddlewares(
-    req: IHttpRequest,
-    res: IHttpResponse,
-    next?: () => void,
-  ): void;
-}
