--- lib/middlewares/middleware.types.ts
@@ -1,7 +1,0 @@
-import { IHttpRequest, IHttpResponse } from "../types";
-
-export type Middleware = (
-  req: IHttpRequest,
-  res: IHttpResponse,
-  next: () => void,
-) => void | Promise<void>;
